+++ CHANGELOG.md
@@ -41,4 +41,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - separated sitelist from main.vue
 - Users can specify which branch to build on Jenkins build page, the default branch is still master
 - Jenkins builds will now check for an existing docker container and delete it if a previous build has failed to clean up
-
+- changed dropdown input suggestions to be compatible with the Tableau inline browser